refactor(monitor): split map hook and storm row parser for quality gates
Move OpenLayers lifecycle into useMonitorOlMap, isolate CSV row parsing,
and configure DeepSource doc coverage skips for new React modules.
